Created HtmlHelper. Extracted create Html to ObservationManagerMenuFile
